refactor: split Config types into config.rs, making main/mod.rs a pure mod (config-3) (#2876)
