+++ .github/workflows/security.yml
@@ -35,11 +35,12 @@ jobs:
       - uses: rustsec/audit-check@69366f33c96575abad1ee0dba8212993eecbe998 # v2.0.0
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
-          ignore: RUSTSEC-2023-0071,RUSTSEC-2024-0436,RUSTSEC-2025-0119,RUSTSEC-2025-0134,RUSTSEC-2026-0002,RUSTSEC-2026-0007,RUSTSEC-2026-0049,RUSTSEC-2026-0097,RUSTSEC-2026-0098,RUSTSEC-2026-0099,RUSTSEC-2026-0104,RUSTSEC-2026-0105,RUSTSEC-2025-0141,RUSTSEC-2026-0145
+          ignore: RUSTSEC-2023-0071,RUSTSEC-2024-0436,RUSTSEC-2025-0119,RUSTSEC-2025-0134,RUSTSEC-2026-0002,RUSTSEC-2026-0007,RUSTSEC-2026-0049,RUSTSEC-2026-0097,RUSTSEC-2026-0098,RUSTSEC-2026-0099,RUSTSEC-2026-0104,RUSTSEC-2026-0105,RUSTSEC-2025-0141,RUSTSEC-2026-0145,RUSTSEC-2026-0179,RUSTSEC-2026-0180,RUSTSEC-2026-0178
           # RUSTSEC-2023-0071 = Marvin Attack: potential key recovery through timing side channels => not used exploitably
           # RUSTSEC-2024-0436 = paste - no longer maintained
           # RUSTSEC-2025-0119 = number_prefix is unmaintained, but we need no change in this lib
           # RUSTSEC-2025-0134 = rustls-pemfile is unmaintained, but only used in development dependencies
+          # RUSTSEC-2025-0141 = bincode unmaintained, is done
           # RUSTSEC-2026-0002 = unsound stack borrow in lru
           # RUSTSEC-2026-0007 = integer overflow in bytes
           # RUSTSEC-2026-0049 = rustls-webpki@0.101.7 CRL matching bug; stuck in aws-smithy-http-client's legacy rustls@0.21 chain, cannot be upgraded without an upstream aws-sdk-s3 release
@@ -48,8 +49,10 @@ jobs:
           # RUSTSEC-2026-0099 = rustls-webpki wildcard name constraints; remaining finding is rustls-webpki@0.101.7 in aws-smithy-http-client's legacy rustls@0.21 chain
           # RUSTSEC-2026-0104 = rustls-webpki CRL parsing panic; only remaining finding is rustls-webpki@0.101.7 in aws-smithy-http-client's legacy rustls@0.21 chain
           # RUSTSEC-2026-0105 = core2 unmaintained
-          # RUSTSEC-2025-0141 = bincode unmaintained, is done
           # RUSTSEC-2026-0145 = astral-tokio-tar header desync, not used exploitably
+          # RUSTSEC-2026-0179 = postgres-protocol dos
+          # RUSTSEC-2026-0180 = postgres-protocol dos
+          # RUSTSEC-2026-0178 = tokio-postgres dos
   dependency-review:
     runs-on: ubuntu-latest
     permissions:
